+++ src/main/java/roomescape/dao/JdbcReservationTimeDao.java
@@ -1,0 +1,49 @@
+package roomescape.dao;
+
+import java.util.List;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.entity.ReservationTime;
+
+@Repository
+public class JdbcReservationTimeDao implements ReservationTimeDao {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
+    private final RowMapper<ReservationTime> rowMapper = (rs, rowNum) -> new ReservationTime(
+            rs.getLong("id"),
+            rs.getTime("start_at").toLocalTime());
+
+    public JdbcReservationTimeDao(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("reservation_time")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT id, start_at FROM reservation_time";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public ReservationTime findById(long id) {
+        String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, rowMapper, id);
+    }
+
+    public long save(ReservationTime reservationTime) {
+        SqlParameterSource parameterSource = new MapSqlParameterSource()
+                .addValue("start_at", reservationTime.getStartAt());
+        return jdbcInsert.executeAndReturnKey(parameterSource).longValue();
+    }
+
+    public int deleteById(long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.update(sql, id);
+    }
+}
